Add a HtmlResponse subclass.
This allows easy creation of responses with string content. No need
to mess with streams.
